+++ apps/web/src/app/components/Menu.ts
@@ -1,3 +1,0 @@
-import { Component } from '@lisergia/core'
-
-export default class Menu extends Component {}