+++ apps/web/stylelint.config.js
@@ -1,3 +1,0 @@
-import config from '@lisergia/config-stylelint'
-
-export default config