+++ packages/config-prettier/index.js
@@ -1,0 +1,21 @@
+module.exports = {
+  printWidth: 120,
+  semi: false,
+  singleQuote: true,
+  tabWidth: 2,
+  trailingComma: 'all',
+  useTabs: false,
+  overrides: [
+    {
+      files: ['*.twig'],
+      options: {
+        parser: 'twig',
+        printWidth: 120,
+        singleQuote: false,
+        tabWidth: 4,
+        useTabs: false,
+      },
+    },
+  ],
+  plugins: ['@zackad/prettier-plugin-twig'],
+}