+++ packages/config-eslint/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @lisergia/config-eslint
 
+## 5.0.0
+
+### Major Changes
+
+- - Add Tempus utilities.
+  - Remove module type from packages.
+  - Add Prettier configuration as dependency.
+  - Move Front End and Back End packages to separate repositories.
+
 ## 4.0.0
 
 ### Major Changes